+++ components/core/Heading.vue
@@ -1,21 +1,28 @@
 <template>
-    <component :is="as" :class="['text-gray-800 mb-3', sizes[as], as === 'h2' ? 'font-semibold' : 'font-medium']">
-        <slot />
-    </component>
+  <component :is="as"
+    :class="twMerge('text-gray-800 mb-3', sizes[as], as === 'h2' ? 'font-semibold' : 'font-medium', $attrs.class as string)">
+    <slot />
+  </component>
 </template>
 
 <script setup lang="ts">
+import { twMerge } from 'tailwind-merge'
+
 interface Props {
-    as?: string;
+  as?: string;
 }
 
 const sizes: Record<string, string> = {
-    h2: "text-3xl",
-    h3: "text-2xl",
-    h4: "text-xl",
+  h2: "text-3xl",
+  h3: "text-2xl",
+  h4: "text-xl",
 };
 
+defineOptions({
+  inheritAttrs: false,
+})
+
 withDefaults(defineProps<Props>(), {
-    as: "h2",
+  as: "h2",
 });
 </script>